requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
Flask-WTF==1.1.1
Flask-Limiter==3.7.0
Flask-SQLAlchemy==3.0.3
marshmallow==3.19.0
pyodbc==4.0.34
Werkzeug==2.3.7
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==9.1.1
openai==0.27.0